pyproject.toml
[project]
name = "psxchatgpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "playwright",
    "llama-cloud-services",
    "llama-index-core",
    "llama-index-readers-file",
    "python-dotenv",
    "llama_index.llms.gemini",
    "llama_index.embeddings.google_genai",
    "llama_index.llms.google_genai",
    "python-dotenv",
    "mcp[cli,core]>=1.6.0",
    "chainlit",
    "anthropic",
    "asyncpg",
]
